Formatting, type migration for IDs, better page titles

+++ iotea/src/main/java/hu/zza/iotea/controller/CommandController.java
@@ -21,13 +21,13 @@ public class CommandController {
   }
 
   @GetMapping("/{id}")
-  public Optional<CommandOutput> getDeviceById(@PathVariable Long id) {
+  public Optional<CommandOutput> getDeviceById(@PathVariable Integer id) {
     return service.getCommandById(id);
   }
 
   @PutMapping("/{id}")
   public CommandOutput postDevice(
-      @PathVariable Long id, @Valid @RequestBody CommandInput command) {
+      @PathVariable Integer id, @Valid @RequestBody CommandInput command) {
     return service.saveCommand(command);
   }
 }
