+++ iotea/src/main/java/hu/zza/iotea/model/response/CommanderProblem.java
@@ -3,7 +3,7 @@ package hu.zza.iotea.model.response;
 import static java.net.HttpURLConnection.HTTP_INTERNAL_ERROR;
 
 public class CommanderProblem extends Problem {
-  public CommanderProblem(Long deviceId, Long commandId, String detail) {
+  public CommanderProblem(Integer deviceId, Integer commandId, String detail) {
     super(
         deviceId,
         commandId,
